Consolidate entry points around a single __main__.py
